Add the simple calculating feature
This feature allows users to click on button to
choose numbers for the first and the second
operands as well as the operator (+, -, *, /).
After clicking "=" button or click another
operator after finishing their expression
(e.g: 1 + 2 and then + 3), display on the screen
with the result of the previous calculus and get
that result as the operand1 for the following one.

In CSS, stylize the calculator screen with an
invisible scrollbar to avoid overflow when the
calculus is too long